+++ includes/modules/testimonials.php
@@ -11,11 +11,9 @@ if ( ! EJO_Base_Module::is_active('testimonials') ) {
     });
 
     //* Remove widget 
-    add_filter( 'ejo_base_unregister_widgets', function($widgets_to_unregister) {
+    add_action( 'widgets_init', function() {
 
         if (! current_user_can( 'manage_options' ) )
-            $widgets_to_unregister[] = 'EJO_Simple_Testimonials_Widget';
-
-        return $widgets_to_unregister;
-    });
+            unregister_widget( 'EJO_Simple_Testimonials_Widget' );
+    }, 11);
 }