+++ CHANGES.md
@@ -1,3 +1,7 @@
+- Simplified unregistering of widgets
+- Improved blog-module and removed contactads module
+- Added ejo_add_extended_theme_support function as a helper
+
 0.1
 - Named plugin EJO Base Module Manager
 - Migrated Module Manager from EJO Base